23.02.24/b:sprint3/step-1/v:0.1.8.0.3
refactoring modals and Ingredient pages;
