+++ src/components/BurgerIngredients/IngredientGroup/IngredientGroupStyles.module.scss
@@ -38,17 +38,20 @@
   color: $theme_white;
 
 
+
   img {
     transition: all ease-in .2s;
     transform: scale(1);
   }
 
   p {
+    transition: all ease-in .2s;
     display: flex;
     align-items: center;
     text-align: center;
 
     svg {
+      transition: all ease-in .2s;
       scale: 1.2;
       padding-left: 6px;
     }
@@ -77,7 +80,6 @@
 }
 
 .ingredient_card:hover {
-  scale: 1.05;
 
   img {
     transform: scale(1.2);
