--- src/components/common/Modal/ModalOverlay/ModalOverlayStyles.module.scss
@@ -1,19 +1,0 @@
-.modal_overlay {
-  position: fixed;
-  z-index: 100;
-  overflow: auto;
-
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-
-  background-color: #000000;
-  opacity: .6;
-}
-
-.modal_error {
-  opacity: 1;
-}
