+++ src/pages/index.ts
@@ -4,6 +4,16 @@ import NotFound404 from "pages/NotFound404.tsx";
 import RegisterPage from "pages/RegisterPage.tsx";
 import ForgotPage from "pages/ForgotPage.tsx";
 import ProfilePage from "pages/ProfilePage.tsx"
-import IngredientDetailsPage from "pages/IngredientDetailsPage/IngredientDetailsPage.tsx"
+import IngredientDetailsModal from "pages/IngredientDetailsPage/IngredientDetailsModal.tsx"
+import IngredientDetailsDefault from "pages/IngredientDetailsPage/IngredientDetailsDefault.tsx"
 
-export {LoginPage, HomePage, NotFound404, ForgotPage, RegisterPage, ProfilePage, IngredientDetailsPage};
+export {
+    LoginPage,
+    HomePage,
+    NotFound404,
+    ForgotPage,
+    RegisterPage,
+    ProfilePage,
+    IngredientDetailsModal,
+    IngredientDetailsDefault
+};
